Fixing the autoload of the rundeck_acl function.
There was an issue with autoloading the function when it was in
lib/puppet/util so I have refactored it to move it into a more
appropriate location under the PuppetX module namespace.
